# api/reservation.go
package api

import (
	"database/sql"
	"errors"
	"fmt"
	"man/db"
	"man/task"
	"net/http"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/google/uuid"
)

// 查询预约详情

type getReservationRequest struct {
	ID uuid.UUID `uri:"id" binding:"required,min=1"`
}

// getReservation 获取预约详情
// @Summary 获取预约详情
// @Description 根据预约 ID 获取预约信息
// @Tags Reservation
// @Accept json
// @Produce json
// @Param id path string true "预约ID"
// @Success 200 {object} db.Reservation
// @Failure 400
// @Failure 404
// @Failure 500
// @Security BearerAuth
// @Router /v1/reservation/{id} [get]
// @Router /admin/reservation/{id} [get]
func (server *Server) getReservation(ctx *gin.Context) {
	var req getReservationRequest
	if err := ctx.ShouldBindUri(&req); err != nil {
		ctx.JSON(http.StatusBadRequest, errorResponse(err))
		return
	}

	seat, err := server.store.GetReservation(ctx, req.ID)
	if err != nil {
		if err == sql.ErrNoRows {
			ctx.JSON(http.StatusNotFound, errorResponse(err))
			return
		}

		ctx.JSON(http.StatusInternalServerError, errorResponse(err))
		return
	}

	ctx.JSON(http.StatusOK, seat)
}

type listReservationRequest struct {
	Page     int32 `form:"page" binding:"required,min=1"`
	PageSize int32 `form:"page_size" binding:"required,min=5,max=50"`
	// 可为空参数
	StartTime *time.Time `form:"start_time" binding:"omitempty"`
	EndTime   *time.Time `form:"end_time" binding:"omitempty"`
	UserID    *int32     `form:"user_id" binding:"omitempty,min=1"`
	SeatID    *int32     `form:"seat_id" binding:"omitempty,min=1"`
	Status    *string    `form:"status" binding:"omitempty"`
}

// listReservation 获取预约列表
// @Summary 条件获取预约列表
// @Description 支持按时间、用户、座位、状态过滤，并分页返回预约信息
// @Tags Reservation
// @Accept json
// @Produce json
// @Param start_time query string false "起始时间 (ISO8601)"
// @Param end_time query string false "结束时间 (ISO8601)"
// @Param user_id query int false "用户ID"
// @Param seat_id query int false "座位ID"
// @Param status query string false "预约状态"
// @Param page query int true "页码，从1开始"
// @Param page_size query int true "每页数量 (5–50)"
// @Success 200 {array} db.Reservation
// @Failure 400
// @Failure 500
// @Security BearerAuth
// @Router /admin/reservation [get]
func (server *Server) listReservation(ctx *gin.Context) {
	var req listReservationRequest
	if err := ctx.ShouldBindQuery(&req); err != nil {
		ctx.JSON(http.StatusBadRequest, errorResponse(err))
		return
	}

	arg := db.ListReservationParams{
		Limit:     req.PageSize,
		Offset:    (req.Page - 1) * req.PageSize,
		StartTime: db.ToNull[sql.NullTime](req.StartTime),
		EndTime:   db.ToNull[sql.NullTime](req.EndTime),
		UserID:    db.ToNull[sql.NullInt32](req.UserID),
		SeatID:    db.ToNull[sql.NullInt32](req.SeatID),
		Status:    db.ToNull[sql.NullString](req.Status),
	}

	seats, err := server.store.ListReservation(ctx, arg)
	if err != nil {
		ctx.JSON(http.StatusInternalServerError, errorResponse(err))
		return
	}

	ctx.JSON(http.StatusOK, seats)
}

type listMyReservationRequest = Pagination

// listMyReservation 获取当前用户的预约列表
// @Summary 获取当前用户的预约记录（分页）
// @Tags Reservation
// @Accept json
// @Produce json
// @Param page query int true "页码，从1开始"
// @Param page_size query int true "每页数量 (5–50)"
// @Success 200 {array} db.Reservation
// @Failure 400
// @Failure 500
// @Security BearerAuth
// @Router /v1/me/reservation [get]
func (server *Server) listMyReservation(ctx *gin.Context) {
	var req listMyReservationRequest
	if err := ctx.ShouldBindQuery(&req); err != nil {
		ctx.JSON(http.StatusBadRequest, errorResponse(err))
		return
	}

	arg := db.ListReservationParams{
		UserID: db.ToNull[sql.NullInt32](getUserID(ctx)),
		Limit:  req.PageSize,
		Offset: (req.Page - 1) * req.PageSize,
	}

	reservations, err := server.store.ListReservation(ctx, arg)
	if err != nil {
		ctx.JSON(http.StatusInternalServerError, errorResponse(err))
		return
	}

	ctx.JSON(http.StatusOK, reservations)
}

type createReservationRequest struct {
	SeatID    int32     `json:"seat_id" binding:"required,min=1"`
	StartTime time.Time `json:"start_time" binding:"required"`
	EndTime   time.Time `json:"end_time" binding:"required"`
}

// createReservation 创建座位预约
// @Summary 创建预约
// @Tags Reservation
// @Accept json
// @Produce json
// @Param data body createReservationRequest true "预约参数"
// @Success 200 {object} db.Reservation
// @Failure 400
// @Failure 404
// @Failure 500
// @Security BearerAuth
// @Router /v1/reservation [post]
func (server *Server) createReservation(ctx *gin.Context) {
	var req createReservationRequest
	if err := ctx.ShouldBindJSON(&req); err != nil {
		ctx.JSON(http.StatusBadRequest, errorResponse(err))
		return
	}

	// 检查座位是否存在且可用
	seat, err := server.store.GetSeat(ctx, req.SeatID)
	if err != nil {
		if err == sql.ErrNoRows {
			ctx.JSON(http.StatusNotFound, errorResponse(errors.New("座位不存在")))
			return
		}
		ctx.JSON(http.StatusInternalServerError, errorResponse(err))
		return
	}
	if !seat.IsAvailable {
		ctx.JSON(http.StatusBadRequest, errorResponse(errors.New("座位不可用")))
		return
	}

	// 检查开始时间是否在当前时间之前，且不超过最大提前时间
	if req.StartTime.Before(time.Now()) || req.StartTime.After(time.Now().Add(server.config.MaxReservationAdvanceDuration)) {
		ctx.JSON(http.StatusBadRequest, errorResponse(fmt.Errorf("开始时间必须在当前时间之后，并且不超过 %s", server.config.MaxReservationAdvanceDuration.String())))
		return
	}

	// 检查时间是否冲突, 保证预约时长大于0，且小于最大预约时长
	if req.StartTime.After(req.EndTime) || req.EndTime.After(req.StartTime.Add(server.config.MaxReservationDuration)) {
		ctx.JSON(http.StatusBadRequest, errorResponse(errors.New("预约时长错误")))
	}

	arg := db.CreateReservationParams{
		SeatID:    req.SeatID,
		StartTime: req.StartTime,
		EndTime:   req.EndTime,
		UserID:    getUserID(ctx),
	}

	reservation, err := server.store.CreateReservation(ctx, arg)
	if err != nil {
		ctx.JSON(http.StatusInternalServerError, errorResponse(err))
		return
	}

	// 提交预约创建任务
	err = server.scheduler.CreateReservationTasks(task.CreateReservationTaskArg{
		ReservationID:        reservation.ID,
		UserID:               reservation.UserID,
		ReservationTime:      reservation.StartTime,
		RemindBeforeDuration: server.config.ReservationRemindBeforeDuration,
		RemindAfterDuration:  server.config.ReservationRemindAfterDuration,
		ViolationDuration:    server.config.ReservationViolationDuration,
	})
	if err != nil {
		ctx.JSON(http.StatusInternalServerError, errorResponse(err))
		return
	}

	ctx.JSON(http.StatusOK, reservation)
}

type deleteReservationRequest struct {
	ID uuid.UUID `uri:"id" binding:"required,min=1"`
}

// deleteReservation 取消预约
// @Summary 取消指定预约
// @Tags Reservation
// @Accept json
// @Produce json
// @Param id path string true "预约ID"
// @Success 200 {string} string "OK"
// @Failure 400
// @Failure 404
// @Failure 500
// @Security BearerAuth
// @Router /v1/reservation/{id} [delete]
func (server *Server) deleteReservation(ctx *gin.Context) {
	var req deleteReservationRequest
	if err := ctx.ShouldBindUri(&req); err != nil {
		ctx.JSON(http.StatusBadRequest, errorResponse(err))
		return
	}

	// 获取预约开始时间
	reservation, err := server.store.GetReservation(ctx, req.ID)
	if err != nil {
		if err == sql.ErrNoRows {
			ctx.JSON(http.StatusNotFound, errorResponse(err))
			return
		}
		ctx.JSON(http.StatusInternalServerError, errorResponse(err))
		return
	}

	// 检查是否可以取消预约
	if reservation.StartTime.Before(time.Now().Add(-server.config.CancellableReservationDuration)) {
		ctx.JSON(http.StatusBadRequest, errorResponse(errors.New("预约已开始，无法取消")))
	}

	err = server.store.DeleteReservation(ctx, req.ID)
	if err != nil {
		ctx.JSON(http.StatusInternalServerError, errorResponse(err))
		return
	}

	ctx.JSON(http.StatusOK, nil)
}

type checkInRequest struct {
	ID uuid.UUID `uri:"id" binding:"required,min=1"`
}

// checkIn 签到
// @Summary 签到已预约座位
// @Tags Reservation
// @Accept json
// @Produce json
// @Param id path string true "预约ID"
// @Success 200 {object} db.Reservation
// @Failure 400
// @Failure 404
// @Failure 500
// @Security BearerAuth
// @Router /v1/reservation/{id}/checkin [post]
func (server *Server) checkIn(ctx *gin.Context) {
	var req checkInRequest
	if err := ctx.ShouldBindUri(&req); err != nil {
		ctx.JSON(http.StatusBadRequest, errorResponse(err))
		return
	}

	arg := db.UpdateReservationStatusParams{
		ID:     req.ID,
		Status: db.ReservationStatusCompleted,
	}

	reservation, err := server.store.UpdateReservationStatus(ctx, arg)
	if err != nil {
		if err == sql.ErrNoRows {
			ctx.JSON(http.StatusNotFound, errorResponse(err))
			return
		}
		ctx.JSON(http.StatusInternalServerError, errorResponse(err))
		return
	}

	ctx.JSON(http.StatusOK, reservation)
}